catch throws in var<mat> tests and move arena_matrix to core
