+++ stan/math/rev/core/arena_matrix.hpp
@@ -1,5 +1,5 @@
-#ifndef STAN_MATH_REV_FUNCTOR_ARENA_MATRIX_HPP
-#define STAN_MATH_REV_FUNCTOR_ARENA_MATRIX_HPP
+#ifndef STAN_MATH_REV_CORE_ARENA_MATRIX_HPP
+#define STAN_MATH_REV_CORE_ARENA_MATRIX_HPP
 
 #include <stan/math/prim/fun/Eigen.hpp>
 #include <stan/math/rev/core.hpp>
