+++ stan/math/rev/fun/from_var_value.hpp
@@ -3,7 +3,7 @@
 
 #include <stan/math/rev/core.hpp>
 #include <stan/math/prim/fun/Eigen.hpp>
-#include <stan/math/rev/functor/arena_matrix.hpp>
+#include <stan/math/rev/core/arena_matrix.hpp>
 #include <stan/math/rev/functor/reverse_pass_callback.hpp>
 
 namespace stan {
