+++ stan/math/rev/fun/to_arena.hpp
@@ -2,7 +2,7 @@
 #define STAN_MATH_REF_FUN_TO_ARENA_HPP
 
 #include <stan/math/rev/meta.hpp>
-#include <stan/math/rev/functor/arena_matrix.hpp>
+#include <stan/math/rev/core/arena_matrix.hpp>
 #include <stan/math/prim/fun/Eigen.hpp>
 #include <vector>
 #include <cstring>
