+++ stan/math/rev/functor.hpp
@@ -1,7 +1,6 @@
 #ifndef STAN_MATH_REV_FUNCTOR_HPP
 #define STAN_MATH_REV_FUNCTOR_HPP
 
-#include <stan/math/rev/functor/arena_matrix.hpp>
 #include <stan/math/rev/functor/adj_jac_apply.hpp>
 #include <stan/math/rev/functor/algebra_solver_fp.hpp>
 #include <stan/math/rev/functor/algebra_solver_powell.hpp>
